[Ops] fix compile warnings

+++ src/core/substitution.cc
@@ -747,7 +747,7 @@ bool GraphXfer::can_match(OpX* srcOp, Op op, Graph* graph)
 {
   if (srcOp->type != op.ptr->type) return false;
   // check num input tensors
-  if (srcOp->inputs.size() != op.ptr->numInputs) return false;
+  if ((int)srcOp->inputs.size() != op.ptr->numInputs) return false;
   // check pmConstraints
   for (size_t i = 0; i < srcOp->pmConstraints.size(); i++) {
     PMConstraint pmc = srcOp->pmConstraints[i];
